pyproject.toml

[project]
name = "immich-mcp"
version = "0.1.0"
description = "Model Context Protocol (MCP) server for Immich API"
authors = [
    { name = "OpenHands", email = "openhands@all-hands.dev" }
]
dependencies = [
    "mcp",
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.1",
    "python-dotenv>=1.0.1",
    "pydantic>=2.0.0",
    "pydantic-settings",
    "pytest",
    "pytest-asyncio",
    "httpx",
    "respx",
    "tenacity",
]
requires-python = ">=3.10"
license = { text = "MIT" }

[project.optional-dependencies]
dev = [
    "pyright",
    "ruff",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]

[tool.pytest.ini_options]
pythonpath = ["."]

[tool.uvicorn.server]
host = "0.0.0.0"
port = 8000
log_level = "info"
